% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-bivariate-analysis.R
\name{blr_bivariate_analysis}
\alias{blr_bivariate_analysis}
\alias{blr_bivariate_analysis.default}
\title{Bivariate Analysis}
\usage{
blr_bivariate_analysis(data, response, ...)

\method{blr_bivariate_analysis}{default}(data, response, ...)
}
\arguments{
\item{data}{a tibble}

\item{response}{response variable}

\item{...}{predictor variables}
}
\value{
a tibble
}
\description{
Bivariate analysis
}
\examples{
blr_bivariate_analysis(hsb2, honcomp, female, prog, race, schtyp)

}
