% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-model-validation.R
\name{blr_confusion_matrix}
\alias{blr_confusion_matrix}
\title{Confusion Matrix}
\usage{
blr_confusion_matrix(model, data, cutoff)
}
\arguments{
\item{model}{an object of class \code{glm}}

\item{data}{a tibble}

\item{cutoff}{cutoff for classification}
}
\value{
table
}
\description{
Confusion matrix
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
            family = binomial(link = 'logit'))

blr_confusion_matrix(model, hsb2, 0.4)
}
