% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-model-fit-stats.R
\name{blr_count_rsq}
\alias{blr_count_rsq}
\title{Count R2}
\usage{
blr_count_rsq(model)
}
\arguments{
\item{model}{An object of class \code{glm}.}
}
\value{
Count r-squared.
}
\description{
Count r-squared.
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
            family = binomial(link = 'logit'))

blr_count_rsq(model)

}
