% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-gains-table.R
\name{blr_decile_capture_rate}
\alias{blr_decile_capture_rate}
\title{Capture Rate by Decile}
\usage{
blr_decile_capture_rate(gains_table, xaxis_title = "Decile",
  yaxis_title = "Event Rate", title = "Capture Rate by Decile",
  bar_color = "blue", text_size = 3.5, text_vjust = -0.3)
}
\arguments{
\item{gains_table}{an object of class \code{blr_gains_table}}

\item{xaxis_title}{x axis title}

\item{yaxis_title}{y axis title}

\item{title}{plot title}

\item{bar_color}{color of the bars}

\item{text_size}{size of the bar labels}

\item{text_vjust}{vertical justification of the bar labels}
}
\description{
Visualize the decile wise capture rate
}
\examples{
model <- glm(honcomp ~ female + read + science, data = blorr::hsb2,
             family = binomial(link = 'logit'))
gt <- blr_gains_table(model)
blr_decile_capture_rate(gt)

}
