% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-gains-table.R
\name{blr_decile_lift_chart}
\alias{blr_decile_lift_chart}
\title{Decile Lift Chart}
\usage{
blr_decile_lift_chart(gains_table, xaxis_title = "Decile",
  yaxis_title = "Decile Mean / Global Mean", title = "Decile Lift Chart",
  bar_color = "blue", text_size = 3.5, text_vjust = -0.3)
}
\arguments{
\item{gains_table}{an object of class \code{blr_gains_table}}

\item{xaxis_title}{x axis title}

\item{yaxis_title}{y axis title}

\item{title}{plot title}

\item{bar_color}{color of the bars}

\item{text_size}{size of the bar labels}

\item{text_vjust}{vertical justification of the bar labels}
}
\description{
Decile wise lift chart
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
             family = binomial(link = 'logit'))
gt <- blr_gains_table(model)
blr_decile_lift_chart(gt)

}
