% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-model-fit-stats.R
\name{blr_effron_rsq}
\alias{blr_effron_rsq}
\title{Effron R2}
\usage{
blr_effron_rsq(model)
}
\arguments{
\item{model}{an object of class \code{glm}}
}
\value{
Effron pseudo r-squared
}
\description{
Effron pseudo r-squared
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
            family = binomial(link = 'logit'))

blr_effron_rsq(model)
}
