% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-forward-selection.R
\name{blr_forward_selection}
\alias{blr_forward_selection}
\alias{blr_forward_selection.default}
\title{Forward Selection}
\usage{
blr_forward_selection(model, enter = 0.3, include = NULL, stop = NULL,
  details = FALSE, ...)

\method{blr_forward_selection}{default}(model, enter = 0.3, include = NULL,
  stop = NULL, details = FALSE, ...)
}
\arguments{
\item{model}{an object of class \code{glm}}

\item{enter}{significance level for entering the variable into the model}

\item{include}{variables that must be included in the model}

\item{details}{logical; if \code{TRUE}, will print the regression result at each step}

\item{...}{other arguments}

\item{steps}{the maximum number of steps to be considered}
}
\value{
a tibble with the following columns

\item{step}{the step in forward selection}
\item{variable}{name of the variable entered}
\item{chisq}{chi square statistic}
\item{pvalue}{p value associated with the chi square statistic}
}
\description{
Build regression model from a set of candidate predictor variables by entering predictors based on
Chi Square statistic, in a stepwise manner until there is no variable left to enter any more.
}
