% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-lorenz-curve.R
\name{blr_gini_index}
\alias{blr_gini_index}
\title{Gini Index}
\usage{
blr_gini_index(model, data = NULL)
}
\arguments{
\item{model}{an object of class \code{glm}}

\item{data}{a tibble or a data.frame}
}
\value{
Gini Index
}
\description{
Gini Index
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
family = binomial(link = 'logit'))

blr_gini_index(model)
}
