% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-hosmer-lemeshow-test.R
\name{blr_hosmer_lemeshow_test}
\alias{blr_hosmer_lemeshow_test}
\title{Hosmer Lemeshow Test}
\usage{
blr_hosmer_lemeshow_test(model, data)
}
\arguments{
\item{model}{an object of class \code{glm}}

\item{data}{a tibble or a data.frame}
}
\description{
Hosmer lemeshow test
}
\examples{
model <- glm(honcomp ~ female + read + science, data = blorr::hsb2,
             family = binomial(link = 'logit'))

blr_hosmer_lemeshow_test(model)
}
