% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-launch-app.R
\name{blr_launch_app}
\alias{blr_launch_app}
\title{Launch shiny app}
\usage{
blr_launch_app()
}
\description{
Launches shiny app for interactive model building.
}
\examples{
\dontrun{
blr_launch_app()
}
}
