% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-linktest.R
\name{blr_linktest}
\alias{blr_linktest}
\title{Model specification error}
\usage{
blr_linktest(model)
}
\arguments{
\item{model}{An object of class \code{glm}.}
}
\value{
An object of class \code{glm}.
}
\description{
Test for model specification error.
}
\examples{

model <- glm(honcomp ~ female + read + science, data = hsb2,
family = binomial(link = 'logit'))

blr_linktest(model)

}
