% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-lorenz-curve.R
\name{blr_lorenz_curve}
\alias{blr_lorenz_curve}
\title{Lorenz Curve}
\usage{
blr_lorenz_curve(model, data = NULL, title = "Lorenz Curve",
  xaxis_title = "Cumulative Population \%",
  yaxis_title = "Cumulative Events \%", diag_line_col = "red",
  lorenz_curve_col = "blue")
}
\arguments{
\item{model}{an object of class \code{glm}}

\item{data}{a tibble or a data.frame}

\item{title}{plot title}

\item{xaxis_title}{x axis title}

\item{yaxis_title}{y axis title}

\item{diag_line_col}{diagonal line color}

\item{lorenz_curve_col}{color of the lorenz curve}
}
\value{
Lorenz Curve
}
\description{
Lorenz curve
}
\examples{
model <- glm(honcomp ~ female + read + science, data = blorr::hsb2,
family = binomial(link = 'logit'))

blr_lorenz_curve(model)
}
