% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-model-fit-stats.R
\name{blr_mcfadden_adj_rsq}
\alias{blr_mcfadden_adj_rsq}
\title{McFadden's adjusted R2}
\usage{
blr_mcfadden_adj_rsq(model)
}
\arguments{
\item{model}{An object of class \code{glm}.}
}
\value{
McFadden's adjusted r-squared.
}
\description{
McFadden's adjusted pseudo r-squared for the model.
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
            family = binomial(link = 'logit'))

blr_mcfadden_adj_rsq(model)

}
\seealso{
Other model fit statistics: \code{\link{blr_adj_count_rsq}},
  \code{\link{blr_cox_snell_rsq}},
  \code{\link{blr_effron_rsq}}, \code{\link{blr_lr_test}},
  \code{\link{blr_mckelvey_zavoina_rsq}},
  \code{\link{blr_model_fit_stats}},
  \code{\link{blr_multi_model_fit_stats}},
  \code{\link{blr_nagelkerke_rsq}}, \code{\link{blr_pairs}}
}
