% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-model-fit-stats.R
\name{blr_mcfadden_rsq}
\alias{blr_mcfadden_rsq}
\title{McFadden's R2}
\usage{
blr_mcfadden_rsq(model)
}
\arguments{
\item{model}{an object of class \code{glm}}
}
\value{
McFadden's r-squared
}
\description{
McFadden's pseudo r-squared for the model.
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
            family = binomial(link = 'logit'))

blr_mcfadden_rsq(model)
}
