% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-model-fit-stats.R
\name{blr_nagelkerke_rsq}
\alias{blr_nagelkerke_rsq}
\title{Cragg-Uhler (Nagelkerke) R2}
\usage{
blr_nagelkerke_rsq(model)
}
\arguments{
\item{model}{an object of class \code{glm}}
}
\value{
Cragg-Uhler (Nagelkerke) R2 pseudo r-squared
}
\description{
Cragg-Uhler (Nagelkerke) R2 pseudo r-squared
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
            family = binomial(link = 'logit'))

blr_nagelkerke_rsq(model)
}
