% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-pairs.R
\name{blr_pairs}
\alias{blr_pairs}
\title{Concordant & Discordant Pairs}
\usage{
blr_pairs(model)
}
\arguments{
\item{model}{an object of class \code{glm}}
}
\value{
a tibble
}
\description{
Association of predicted probabilities and observed responses
}
\examples{
model <- glm(honcomp ~ female + read + science, data = blorr::hsb2,
family = binomial(link = 'logit'))

blr_pairs(model)
}
