% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-residual-diagnostics.R
\name{blr_plot_diag_fit}
\alias{blr_plot_diag_fit}
\title{Fitted Values Diagnostics Plot}
\usage{
blr_plot_diag_fit(model)
}
\arguments{
\item{model}{an object of class \code{glm}}
}
\value{
a panel of fitted values diagnostic plots
}
\description{
Diagnostic plots for fitted values
}
\examples{
model <- glm(honcomp ~ female + read + science, data = blorr::hsb2,
family = binomial(link = 'logit'))

blr_plot_diag_fit(model)
}
