% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-residual-diagnostics.R
\name{blr_plot_diag_leverage}
\alias{blr_plot_diag_leverage}
\title{Leverage Diagnostics Plot}
\usage{
blr_plot_diag_leverage(model)
}
\arguments{
\item{model}{an object of class \code{glm}}
}
\value{
a panel of diagnostic plots for leverage
}
\description{
Diagnostic plots for leverage
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
family = binomial(link = 'logit'))

blr_plot_diag_leverage(model)
}
