% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-plots-data.R
\name{blr_prep_dcrate_data}
\alias{blr_prep_dcrate_data}
\title{Decile capture rate data}
\usage{
blr_prep_dcrate_data(gains_table)
}
\arguments{
\item{gains_table}{An object of clas \code{blr_gains_table}}
}
\description{
Data for generating decile capture rate.
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
             family = binomial(link = 'logit'))
gt <- blr_gains_table(model)
blr_prep_dcrate_data(gt)

}
