% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-regress.R
\name{blr_regress}
\alias{blr_regress}
\alias{blr_regress.glm}
\title{Binary Logistic Regression}
\usage{
blr_regress(object, ...)

\method{blr_regress}{glm}(object, ...)
}
\arguments{
\item{object}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted or class
\code{glm}}

\item{...}{other inputs}
}
\description{
Binary logistic regression
}
\examples{
blr_regress(object = honcomp ~ female + read + science, data = hsb2)
}
