% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-plots.R
\name{blr_residual_diagnostics}
\alias{blr_residual_diagnostics}
\title{Residual diagnostics}
\usage{
blr_residual_diagnostics(model)
}
\arguments{
\item{model}{An object of class \code{glm}.}
}
\value{
C, CBAR, DIFDEV and DIFCHISQ.
}
\description{
Diagnostics for confidence interval displacement and detecting ill fitted
observations.
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
family = binomial(link = 'logit'))

blr_residual_diagnostics(model)

}
