% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-roc-curve.R
\name{blr_roc_curve}
\alias{blr_roc_curve}
\title{ROC Curve}
\usage{
blr_roc_curve(gains_table, title = "ROC Curve",
  xaxis_title = "1 - Specificity", yaxis_title = "Sensitivity",
  roc_curve_col = "blue", point_shape = 18, point_fill = "blue",
  point_color = "blue", plot_title_justify = 0.5)
}
\arguments{
\item{gains_table}{an object of class \code{blr_gains_table}}
}
\description{
ROC Curve
}
\examples{
model <- glm(honcomp ~ female + read + science, data = blorr::hsb2,
             family = binomial(link = 'logit'))
k <- blr_gains_table(model, hsb2)
blr_roc_curve(k)
}
