% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-bivariate-analysis.R
\name{blr_segment}
\alias{blr_segment}
\alias{blr_segment.default}
\title{Response by Segments}
\usage{
blr_segment(data, response, predictor)

\method{blr_segment}{default}(data, response, predictor)
}
\arguments{
\item{data}{a tibble}

\item{response}{response variable}

\item{predictor}{predictor variable}
}
\value{
a tibble
}
\description{
Response by segements
}
\examples{
blr_segment(hsb2, honcomp, prog)
}
