% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-bivariate-analysis.R
\name{blr_segment_twoway}
\alias{blr_segment_twoway}
\alias{blr_segment_twoway.default}
\title{Two way event rate}
\usage{
blr_segment_twoway(data, response, variable_1, variable_2)

\method{blr_segment_twoway}{default}(data, response, variable_1, variable_2)
}
\arguments{
\item{data}{A \code{tibble} or \code{data.frame}.}

\item{response}{Response variable; column in \code{data}.}

\item{variable_1}{Column in \code{data}.}

\item{variable_2}{Column in \code{data}.}
}
\value{
A tibble.
}
\description{
Event rate across two qualitative variables.
}
\examples{
blr_segment_twoway(hsb2, honcomp, prog, female)

}
\seealso{
Other bivariate analysis procedures: 
\code{\link{blr_bivariate_analysis}()},
\code{\link{blr_segment_dist}()},
\code{\link{blr_segment}()},
\code{\link{blr_woe_iv_stats}()},
\code{\link{blr_woe_iv}()}
}
\concept{bivariate analysis procedures}
