% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-stepwise-selection.R
\name{blr_stepwise_selection}
\alias{blr_stepwise_selection}
\alias{plot.blr_stepwise_selection}
\title{Stepwise AIC selection}
\usage{
blr_stepwise_selection(model, details = FALSE, ...)

\method{plot}{blr_stepwise_selection}(x, ...)
}
\arguments{
\item{model}{An object of class \code{lm}.}

\item{details}{Logical; if \code{TRUE}, details of variable selection will be
printed on screen.}

\item{...}{Other arguments.}

\item{x}{An object of class \code{blr_stepwise_selection}.}
}
\value{
\code{blr_stepwise_selection} returns an object of class \code{"blr_stepwise_selection"}.
An object of class \code{"blr_stepwise_selection"} is a list containing the
following components:

\item{candidates}{candidate predictor variables}
\item{predictors}{variables added/removed from the model}
\item{method}{addition/deletion}
\item{aics}{akaike information criteria}
\item{bics}{bayesian information criteria}
\item{devs}{deviances}
\item{steps}{total number of steps}
}
\description{
Build regression model from a set of candidate predictor variables by
entering and removing predictors based on akaike information criterion, in a
stepwise manner until there is no variable left to enter or remove any more.
}
\examples{
\dontrun{
model <- glm(y ~ ., data = stepwise)

# selection summary
blr_stepwise_selection(model)

# print details at each step
blr_stepwise_selection(model, details = TRUE)

# plot
plot(blr_stepwise_selection(model))
}

}
\references{
Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.
}
\seealso{
Other variable selection procedures: \code{\link{blr_backward_elimination}},
  \code{\link{blr_forward_selection}}
}
