% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-woe-iv.R
\name{blr_woe_iv}
\alias{blr_woe_iv}
\title{Weight of Evidence & Information Value}
\usage{
blr_woe_iv(data, predictor, response)
}
\arguments{
\item{data}{a tibble}

\item{predictor}{name of the predictor}

\item{response}{name of the response variable}
}
\value{
a tibble
}
\description{
Weight of evidence and information value
}
