% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-woe-iv.R
\name{blr_woe_iv_stats}
\alias{blr_woe_iv_stats}
\title{Multi Variable WOE & IV}
\usage{
blr_woe_iv_stats(data, response, ...)
}
\arguments{
\item{data}{a data.frame or a tibble}

\item{response}{response variables}

\item{...}{predictor variables}
}
\description{
Prints WOE & IV for multiple variables
}
\examples{
blr_woe_iv_stats(hsb2, honcomp, prog, race, female, schtyp)
}
