% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-model-fit-stats.R
\name{mcfadden_adj_rsq}
\alias{mcfadden_adj_rsq}
\title{McFadden's Adjusted R2}
\usage{
mcfadden_adj_rsq(model)
}
\arguments{
\item{model}{an object of class \code{glm}}
}
\value{
McFadden's adjusted r-squared
}
\description{
McFadden's adjusted pseudo r-squared for the model.
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
            family = binomial(link = 'logit'))

mcfadden_adj_rsq(model)
}
