% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-data-stepwise.R
\docType{data}
\name{stepwise}
\alias{stepwise}
\title{Dummy Data Set}
\format{An object of class \code{data.frame} with 20000 rows and 7 columns.}
\usage{
stepwise
}
\description{
Dummy Data Set
}
\keyword{datasets}
