% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-blorr.R
\docType{package}
\name{blorr}
\alias{blorr}
\alias{blorr-package}
\title{\code{blorr} package}
\description{
Tools for developing binary logistic regression models
}
\details{
See the README on
\href{https://github.com/rsquaredacademy/blorr}{GitHub}
}
