% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-model-fit-stats.R
\name{blr_adj_count_rsq}
\alias{blr_adj_count_rsq}
\title{Adjusted Count R2}
\usage{
blr_adj_count_rsq(model)
}
\arguments{
\item{model}{an object of class \code{glm}}
}
\value{
Adjusted Count r-squared
}
\description{
Adjusted Count r-squared
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
            family = binomial(link = 'logit'))

blr_adj_count_rsq(model)
}
