% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-backward-elimination.R
\name{blr_backward_elimination}
\alias{blr_backward_elimination}
\alias{blr_backward_elimination.default}
\alias{plot.blr_backward_elimination}
\title{Stepwise AIC backward elimination}
\usage{
blr_backward_elimination(model, details = FALSE, ...)

\method{blr_backward_elimination}{default}(model, details = FALSE, ...)

\method{plot}{blr_backward_elimination}(x, ...)
}
\arguments{
\item{model}{An object of class \code{glm}; the model should include all
candidate predictor variables.}

\item{details}{Logical; if \code{TRUE}, will print the regression result at
each step.}

\item{...}{Other arguments.}

\item{x}{An object of class \code{blr_backward_elimination}.}
}
\value{
\code{blr_backward_elimination} returns an object of class
\code{"blr_backward_elimination"}. An object of class
\code{"blr_backward_elimination"} is a list containing the following components:

\item{candidates}{candidate predictor variables}
\item{steps}{total number of steps}
\item{predictors}{variables removed from the model}
\item{aics}{akaike information criteria}
\item{bics}{bayesian information criteria}
\item{devs}{deviances}
}
\description{
Build regression model from a set of candidate predictor variables by
removing predictors based on akaike information criterion, in a stepwise
manner until there is no variable left to remove any more.
}
\examples{
\dontrun{
model <- glm(honcomp ~ female + read + science + math + prog + socst,
data = hsb2, family = binomial(link = 'logit'))

# elimination summary
blr_backward_elimination(model)

# print details of each step
blr_backward_elimination(model, details = TRUE)

# plot
plot(blr_backward_elimination(model))
}

}
\references{
Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.
}
\seealso{
Other variable selection procedures: \code{\link{blr_forward_selection}},
  \code{\link{blr_stepwise_selection}}
}
