% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-model-validation.R
\name{blr_confusion_matrix}
\alias{blr_confusion_matrix}
\title{Confusion Matrix}
\usage{
blr_confusion_matrix(model, cutoff = 0.5, data = NULL)
}
\arguments{
\item{model}{an object of class \code{glm}}

\item{cutoff}{cutoff for classification}

\item{data}{a tibble}
}
\value{
Confusion Matrix
}
\description{
Confusion matrix
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
            family = binomial(link = 'logit'))

blr_confusion_matrix(model, cutoff = 0.4)
}
