% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-model-fit-stats.R
\name{blr_cox_snell_rsq}
\alias{blr_cox_snell_rsq}
\title{Cox Snell R2}
\usage{
blr_cox_snell_rsq(model)
}
\arguments{
\item{model}{an object of class \code{glm}}
}
\value{
Cox Snell pseudo r-squared
}
\description{
Cox Snell pseudo r-squared
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
            family = binomial(link = 'logit'))

blr_cox_snell_rsq(model)
}
