% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-forward-selection.R
\name{blr_forward_selection}
\alias{blr_forward_selection}
\alias{blr_forward_selection.default}
\alias{plot.blr_forward_selection}
\title{Forward Selection}
\usage{
blr_forward_selection(model, details = FALSE, ...)

\method{blr_forward_selection}{default}(model, details = FALSE, ...)

\method{plot}{blr_forward_selection}(x, ...)
}
\arguments{
\item{model}{an object of class \code{glm}}

\item{details}{logical; if \code{TRUE}, will print the regression result at each step}

\item{...}{other arguments}

\item{x}{An object of class \code{blr_forward_selection}.}
}
\value{
\code{blr_forward_selection} returns an object of class
\code{"blr_forward_selection"}. An object of class
\code{"blr_forward_selection"} is a list containing the following components:

\item{candidates}{candidate predictor variables}
\item{steps}{total number of steps}
\item{predictors}{variables entered into the model}
\item{aics}{akaike information criteria}
\item{bics}{bayesian information criteria}
\item{devs}{deviances}
}
\description{
Build regression model from a set of candidate predictor variables by entering predictors based on
Chi Square statistic, in a stepwise manner until there is no variable left to enter any more.
}
\examples{
\dontrun{
model <- glm(honcomp ~ female + read + science, data = hsb2,
family = binomial(link = 'logit'))

# selection summary
blr_forward_selection(model)

# print details of each step
blr_forward_selection(model, details = TRUE)

# plot
plot(blr_forward_selection(model))
}
}
\references{
Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.
}
