% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-forward-selection.R
\name{blr_forward_selection}
\alias{blr_forward_selection}
\alias{blr_forward_selection.default}
\title{Forward Selection}
\usage{
blr_forward_selection(model, details = FALSE, ...)

\method{blr_forward_selection}{default}(model, details = FALSE, ...)
}
\arguments{
\item{model}{an object of class \code{glm}}

\item{details}{logical; if \code{TRUE}, will print the regression result at each step}

\item{...}{other arguments}
}
\value{
a tibble with the following columns

\item{step}{the step in forward selection}
\item{variable}{name of the variable entered}
\item{chisq}{chi square statistic}
\item{pvalue}{p value associated with the chi square statistic}
}
\description{
Build regression model from a set of candidate predictor variables by entering predictors based on
Chi Square statistic, in a stepwise manner until there is no variable left to enter any more.
}
\examples{
\dontrun{
model <- glm(honcomp ~ female + read + science, data = blorr::hsb2,
family = binomial(link = 'logit'))

blr_forward_selection(model)
blr_forward_selection(model, details = TRUE)
}
}
