% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-gains-table.R
\name{blr_gains_table}
\alias{blr_gains_table}
\title{Gains Table & Lift Curve}
\usage{
blr_gains_table(model, data = NULL)
}
\arguments{
\item{model}{an object of class \code{glm}}

\item{data}{a tibble}
}
\value{
a tibble
}
\description{
Gains table
}
\examples{
model <- glm(honcomp ~ female + read + science, data = blorr::hsb2,
             family = binomial(link = 'logit'))
blr_gains_table(model)
}
