% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-lorenz-curve.R
\name{blr_gini_index}
\alias{blr_gini_index}
\title{Gini index}
\usage{
blr_gini_index(model, data = NULL)
}
\arguments{
\item{model}{An object of class \code{glm}.}

\item{data}{A \code{tibble} or \code{data.frame}.}
}
\value{
Gini index.
}
\description{
Gini index.
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
family = binomial(link = 'logit'))

blr_gini_index(model)

}
\seealso{
Other model validation techniques: \code{\link{blr_confusion_matrix}},
  \code{\link{blr_decile_capture_rate}},
  \code{\link{blr_decile_lift_chart}},
  \code{\link{blr_gains_table}},
  \code{\link{blr_ks_chart}},
  \code{\link{blr_lorenz_curve}},
  \code{\link{blr_roc_curve}},
  \code{\link{blr_test_hosmer_lemeshow}}
}
