% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-gains-table.R
\name{blr_ks_chart}
\alias{blr_ks_chart}
\title{KS Chart}
\usage{
blr_ks_chart(gains_table, title = "KS Chart", yaxis_title = " ",
  xaxis_title = "Cumulative Population \%", ks_line_color = "black")
}
\arguments{
\item{gains_table}{an object of class \code{blr_gains_table}}

\item{title}{plot title}

\item{yaxis_title}{y axis title}

\item{xaxis_title}{x axis title}

\item{ks_line_color}{color of the line indicating maximum ks}
}
\description{
Kolmogorov Smirnov statisitcs graph
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
             family = binomial(link = 'logit'))
gt <- blr_gains_table(model)
blr_ks_chart(gt)

}
