% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-lrtest.R
\name{blr_lr_test}
\alias{blr_lr_test}
\title{Likelihood Ratio Test}
\usage{
blr_lr_test(full_model, reduced_model)
}
\arguments{
\item{full_model}{an object of class \code{glm}; model with all predictors}

\item{reduced_model}{an object of class \code{glm}; nested model. Optional if
you are comparing the \code{full_model} with an intercept only model.}
}
\value{
two tibbles with model information and test results
}
\description{
Performs the likelihood ratio test for full and reduced model.
}
\examples{
# compare full model with intercept only model
hsb <- descriptr::hsb

create honcomp variable
hsb \%<>\%
 mutate(honcomp = if_else(write >= 60, 1, 0))

regression
model_1 <- glm(honcomp ~ female + read + science, data = hsb,
            family = binomial(link = 'logit'))

blr_ls_test(model_1)

# compare full model with nested model
model_2 <- glm(honcomp ~ female + read, data = hsb,
            family = binomial(link = 'logit'))

blr_ls_test(model_1, model_2)
}
\seealso{
\code{\link[lmtest]{lrtest}}
}
