% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-model-fit-stats.R
\name{blr_model_fit_stats}
\alias{blr_model_fit_stats}
\title{Model Fit Statistics}
\usage{
blr_model_fit_stats(model, ...)
}
\arguments{
\item{model}{an object of class \code{glm}}

\item{...}{other inputs}
}
\description{
Model Fit Statistics
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
            family = binomial(link = 'logit'))

blr_model_fit_stats(model)
}
