% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-model-fit-stats.R
\name{blr_model_fit_stats}
\alias{blr_model_fit_stats}
\title{Model fit statistics}
\usage{
blr_model_fit_stats(model, ...)
}
\arguments{
\item{model}{An object of class \code{glm}.}

\item{...}{Other inputs.}
}
\description{
Model fit statistics.
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
            family = binomial(link = 'logit'))

blr_model_fit_stats(model)

}
\seealso{
Other model fit statistics: \code{\link{blr_adj_count_rsq}},
  \code{\link{blr_cox_snell_rsq}},
  \code{\link{blr_effron_rsq}}, \code{\link{blr_lr_test}},
  \code{\link{blr_mcfadden_adj_rsq}},
  \code{\link{blr_mckelvey_zavoina_rsq}},
  \code{\link{blr_multi_model_fit_stats}},
  \code{\link{blr_nagelkerke_rsq}}, \code{\link{blr_pairs}}
}
