% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-multi-model-fit-stats.R
\name{blr_multi_model_fit_stats}
\alias{blr_multi_model_fit_stats}
\alias{blr_multi_model_fit_stats.default}
\title{Multi model fit statistics}
\usage{
blr_multi_model_fit_stats(model, ...)

\method{blr_multi_model_fit_stats}{default}(model, ...)
}
\arguments{
\item{model}{An object of class \code{glm}.}

\item{...}{Objects of class \code{glm}.}
}
\value{
A tibble.
}
\description{
Measures of model fit statistics for multiple models.
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
family = binomial(link = 'logit'))

model2 <- glm(honcomp ~ female + read + math, data = hsb2,
family = binomial(link = 'logit'))

blr_multi_model_fit_stats(model, model2)

}
\seealso{
Other model fit statistics: \code{\link{blr_adj_count_rsq}},
  \code{\link{blr_cox_snell_rsq}},
  \code{\link{blr_effron_rsq}}, \code{\link{blr_lr_test}},
  \code{\link{blr_mcfadden_adj_rsq}},
  \code{\link{blr_mckelvey_zavoina_rsq}},
  \code{\link{blr_model_fit_stats}},
  \code{\link{blr_nagelkerke_rsq}}, \code{\link{blr_pairs}}
}
