% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-multi-model-fit-stats.R
\name{blr_multi_model_fit_stats}
\alias{blr_multi_model_fit_stats}
\alias{blr_multi_model_fit_stats.default}
\title{Multi Model Fit Statistics}
\usage{
blr_multi_model_fit_stats(model, ...)

\method{blr_multi_model_fit_stats}{default}(model, ...)
}
\arguments{
\item{model}{an object of class \code{glm}}

\item{...}{objects of class \code{glm}}
}
\value{
a tibble
}
\description{
Measures of model fit statistics for multiple models
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
family = binomial(link = 'logit'))

model2 <- glm(honcomp ~ female + read + math, data = hsb2,
family = binomial(link = 'logit'))

blr_multi_model_fit_stats(model, model2)

}
