% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-pairs.R
\name{blr_pairs}
\alias{blr_pairs}
\title{Concordant & discordant pairs}
\usage{
blr_pairs(model)
}
\arguments{
\item{model}{An object of class \code{glm}.}
}
\value{
A tibble.
}
\description{
Association of predicted probabilities and observed responses.
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
family = binomial(link = 'logit'))

blr_pairs(model)

}
\seealso{
Other model fit statistics: \code{\link{blr_adj_count_rsq}},
  \code{\link{blr_cox_snell_rsq}},
  \code{\link{blr_effron_rsq}}, \code{\link{blr_lr_test}},
  \code{\link{blr_mcfadden_adj_rsq}},
  \code{\link{blr_mckelvey_zavoina_rsq}},
  \code{\link{blr_model_fit_stats}},
  \code{\link{blr_multi_model_fit_stats}},
  \code{\link{blr_nagelkerke_rsq}}
}
