% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-plots.R
\name{blr_plot_deviance_residual}
\alias{blr_plot_deviance_residual}
\title{Deviance Residual Values}
\usage{
blr_plot_deviance_residual(model, point_color = "blue",
  title = "Deviance Residuals Plot", xaxis_title = "id",
  yaxis_title = "Deviance Residuals")
}
\arguments{
\item{model}{an object of class \code{glm}}

\item{point_color}{color of the points}

\item{title}{title of the plot}

\item{xaxis_title}{x axis label}

\item{yaxis_title}{y axis label}
}
\description{
Deviance residuals plot
}
\examples{
model <- glm(honcomp ~ female + read + science, data = blorr::hsb2,
family = binomial(link = 'logit'))

blr_plot_deviance_residual(model)

}
