% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-plots.R
\name{blr_plot_diag_cbar}
\alias{blr_plot_diag_cbar}
\title{CI Displacement CBAR Plot}
\usage{
blr_plot_diag_cbar(model, point_color = "blue",
  title = "CI Displacement CBAR Plot", xaxis_title = "id",
  yaxis_title = "CI Displacement CBAR")
}
\arguments{
\item{model}{an object of class \code{glm}}

\item{point_color}{color of the points}

\item{title}{title of the plot}

\item{xaxis_title}{x axis label}

\item{yaxis_title}{y axis label}
}
\description{
Confidence interval displacement diagnostics CBAR plot
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
family = binomial(link = 'logit'))

blr_plot_diag_cbar(model)

}
