% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-plots.R
\name{blr_plot_difdev_leverage}
\alias{blr_plot_difdev_leverage}
\title{Delta Deviance vs Leverage Plot}
\usage{
blr_plot_difdev_leverage(model, point_color = "blue",
  title = "Delta Deviance vs Leverage Plot", xaxis_title = "Leverage",
  yaxis_title = "Delta Deviance")
}
\arguments{
\item{model}{an object of class \code{glm}}

\item{point_color}{color of the points}

\item{title}{title of the plot}

\item{xaxis_title}{x axis label}

\item{yaxis_title}{y axis label}
}
\description{
Delta deviance vs leverage plot
}
\examples{
model <- glm(honcomp ~ female + read + science, data = hsb2,
family = binomial(link = 'logit'))

blr_plot_difdev_leverage(model)

}
