% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-regress.R
\name{blr_regress}
\alias{blr_regress}
\alias{blr_regress.glm}
\title{Ordinary Least Squares Regression}
\usage{
blr_regress(formula, ...)

\method{blr_regress}{glm}(object, ...)
}
\arguments{
\item{...}{other inputs}

\item{object}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted or class
\code{glm}}
}
\description{
Ordinary Least Squares Regression
}
\examples{
blr_regress(formula = honcomp ~ female + read + science, data = hsb2)
}
