% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-bivariate-analysis.R
\name{blr_segment}
\alias{blr_segment}
\alias{blr_segment.default}
\title{Event distribution}
\usage{
blr_segment(data, response, predictor)

\method{blr_segment}{default}(data, response, predictor)
}
\arguments{
\item{data}{A \code{tibble} or \code{data.frame}.}

\item{response}{Response variable; column in \code{data}.}

\item{predictor}{Predictor variable; column in \code{data}.}
}
\value{
A tibble.
}
\description{
Event distribution by segements/levels of a qualitative variable.
}
\examples{
blr_segment(hsb2, honcomp, prog)

}
\seealso{
Other bivariate analysis procedures: \code{\link{blr_bivariate_analysis}},
  \code{\link{blr_segment_dist}},
  \code{\link{blr_segment_twoway}},
  \code{\link{blr_woe_iv_stats}}, \code{\link{blr_woe_iv}}
}
