% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-bivariate-analysis.R
\name{blr_segment_dist}
\alias{blr_segment_dist}
\title{Response Distribution by Segments}
\usage{
blr_segment_dist(data, response, predictor)
}
\arguments{
\item{data}{a tibble}

\item{response}{response variable}

\item{predictor}{predictor variable}

\item{x}{an object of class \code{blr_segment_dist}}

\item{title}{plot title}

\item{xaxis_title}{x axis title}

\item{yaxis_title}{y axis title}

\item{sec_yaxis_title}{secondary y axis title}

\item{bar_color}{color of the bar}

\item{line_color}{line color}

\item{...}{other inputs}
}
\value{
a tibble
}
\description{
Distribution of response by segements
}
\examples{
k <- blr_segment_dist(hsb2, honcomp, prog)
plot(k)
}
