% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-bivariate-analysis.R
\name{blr_twoway_segment}
\alias{blr_twoway_segment}
\alias{blr_twoway_segment.default}
\title{Response Distribution by Multiple Segments}
\usage{
blr_twoway_segment(data, response, variable_1, variable_2)

\method{blr_twoway_segment}{default}(data, response, variable_1, variable_2)
}
\arguments{
\item{data}{a tibble}

\item{response}{response variable}

\item{variable_1}{column in \code{data}}

\item{variable_2}{column in \code{data}}
}
\value{
a tibble
}
\description{
Response distribution by multiple segements
}
\examples{
blr_twoway_segment(hsb2, honcomp, prog, female)
}
