% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-woe-iv.R
\name{blr_woe_iv}
\alias{blr_woe_iv}
\title{Weight of Evidence & Information Value}
\usage{
blr_woe_iv(data, predictor, response, digits = 2)
}
\arguments{
\item{data}{a tibble}

\item{predictor}{name of the predictor}

\item{response}{name of the response variable}

\item{digits}{number of decimal digits to round off}
}
\value{
a tibble
}
\description{
Weight of evidence and information value
}
\examples{
blr_woe_iv(hsb2, female, honcomp)
}
