% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-woe-iv.R
\name{blr_woe_iv_stats}
\alias{blr_woe_iv_stats}
\title{Multi variable WOE & IV}
\usage{
blr_woe_iv_stats(data, response, ...)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble}.}

\item{response}{Response variable; column in \code{data}.}

\item{...}{Predictor variables; column in \code{data}.}
}
\description{
Prints weight of evidence and information value for multiple variables. 
  Currently avialable for categorical predictors only.
}
\examples{
blr_woe_iv_stats(hsb2, honcomp, prog, race, female, schtyp)

}
\seealso{
Other bivariate analysis procedures: \code{\link{blr_bivariate_analysis}},
  \code{\link{blr_segment_dist}},
  \code{\link{blr_segment_twoway}},
  \code{\link{blr_segment}}, \code{\link{blr_woe_iv}}
}
\concept{bivariate analysis procedures}
